% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_horizon_df_2.R
\name{build_horizon_df_2}
\alias{build_horizon_df_2}
\title{Compute group for stat_horizon}
\usage{
build_horizon_df_2(data, origin, horizonscale, rm.outliers, reverse, mirror)
}
\arguments{
\item{data}{A data frame.}

\item{origin}{Origin of the horizon plot. It can either be a string, namely
\code{"midpoint"} (the default), \code{"median"},
\code{"mean"}, \code{"min"} or \code{"quantiles"}, or a user-specified number.}

\item{horizonscale}{Cutpoints of the horizon plot. It can either be an integer
specifying the number of ranges (default is \code{6}),
or a user-supplied numeric vector with the cutpoints defining the different ranges.}

\item{rm.outliers}{If \code{TRUE}, all the values below \code{quantile(y, 0.25)-1.5*IQR(y)}
and above \code{quantile(y, 0.75)+1.5*IQR(y)} are excluded from the origin and cutpoint
calculations (default is \code{FALSE}). @param reverse If \code{TRUE}, the horizon peaks
for the values below the origin are reversed (default is \code{FALSE}).}

\item{reverse}{IF \code{TRUE}, the horizon peaks for all the values below the origin
are reversed (default is \code{FALSE}).}

\item{mirror}{If \code{TRUE}, the horizon peaks for all the values are reversed
(default is \code{FALSE}).}
}
\value{
A data frame transformed for a horizon plot.
}
\description{
Compute group for stat_horizon
}
\keyword{internal}
