% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{climate_CPH}
\alias{climate_CPH}
\title{Average temperature in Copenhagen}
\format{
A data frame with 9,132 rows and 9 variables:
\describe{
  \item{Region}{geographic region}
  \item{Country}{country}
  \item{State}{state or territory}
  \item{City}{city or town}
  \item{Month}{month}
  \item{Day}{day}
  \item{Year}{year}
  \item{AvgTemperature}{average temperature in Fahrenheit}
  \item{date_mine}{data in yyy-mm-dd format}
}
}
\source{
\url{https://www.kaggle.com/sudalairajkumar/daily-temperature-of-major-cities}
}
\usage{
climate_CPH
}
\description{
A dataset containing the average temperature in degrees Celsius
for Copenhagen between the year 1995 and 2019.
}
\keyword{datasets}
