% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{climate_US}
\alias{climate_US}
\title{Average temperature in major cities of the US}
\format{
A data frame with 57,828 rows and 9 variables:
\describe{
  \item{Region}{geographic region}
  \item{Country}{country}
  \item{State}{state or territory}
  \item{City}{city or town}
  \item{Month}{month}
  \item{Day}{day}
  \item{Year}{year}
  \item{AvgTemperature}{average temperature in Fahrenheit}
  \item{date_mine}{data in yyy-mm-dd format}
}
}
\source{
\url{https://www.kaggle.com/sudalairajkumar/daily-temperature-of-major-cities, https://benschmidt.org/2014/06/05/optimally-ordering-geographical-entities-in-linear-space/}
}
\usage{
climate_US
}
\description{
A dataset containing the average temperature in Fahrenheit for major cities in the US for the year 2000.
}
\keyword{datasets}
