% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{COVID}
\alias{COVID}
\title{Distribution of COVID-19 cases in Asia}
\format{
A data frame with 12,695 rows and 3 variables:
\describe{
  \item{date_mine}{date of the measurement in yyyy-mm-dd format}
  \item{y}{standardized number of cases}
  \item{countriesAndTerritories}{countries and territories in Asia}
}
}
\source{
\url{https://www.ecdc.europa.eu/en/publications-data/download-todays-data-geographic-distribution-covid-19-cases-worldwide}
}
\usage{
COVID
}
\description{
A dataset containing the geographic distribution of COVID-19 cases in Asia during 2020.
}
\keyword{datasets}
