% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_horizon.R
\name{geom_horizon}
\alias{geom_horizon}
\title{Add together two numbers}
\usage{
geom_horizon(
  mapping = NULL,
  data = NULL,
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = TRUE,
  inherit.aes = TRUE,
  origin = "midpoint",
  horizonscale = 6,
  rm.outliers = FALSE,
  reverse = FALSE
)
}
\description{
Add together two numbers
}
