% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_horizon.R
\docType{data}
\name{GeomHorizon}
\alias{GeomHorizon}
\title{Add together two numbers}
\format{
An object of class \code{GeomHorizon} (inherits from \code{GeomRibbon}, \code{Geom}, \code{ggproto}, \code{gg}) of length 2.
}
\usage{
GeomHorizon
}
\description{
Add together two numbers
}
\keyword{datasets}
