% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_hcl.R
\name{hcl_generic}
\alias{hcl_generic}
\title{General color function}
\usage{
hcl_generic(reverse, palette, ...)
}
\arguments{
\item{reverse}{Logical, indicating whether the palette should be inverted.}

\item{palette}{A character string, indicating the palette to be used.}
}
\value{
A function, that, given an integer, returns a character vector
of colors of the specified palette.
}
\description{
General color function
}
\keyword{internal}
