% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ori_sca.R
\name{is.wholenumber}
\alias{is.wholenumber}
\title{Check if a variable is a whole number}
\usage{
is.wholenumber(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{A number.}

\item{tol}{Toleration threshold.}
}
\description{
Check if a variable is a whole number
}
\keyword{internal}
