% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rmsk}
\alias{rmsk}
\title{Distribution of repeats along the human genome}
\format{
A data frame with 30,885 rows and 4 variables:
\describe{
  \item{genoName}{chromosome name}
  \item{bin}{starting coordinate of window}
  \item{bin_2}{end coordinate of window}
  \item{p_repeat}{percentage of repeats}
}
}
\source{
\url{https://genome.ucsc.edu/cgi-bin/hgTrackUi?g=rmsk}
}
\usage{
rmsk
}
\description{
A dataset containing the percentage of simple repeats
in 100 kb windows along the human genome (hg38).
}
\keyword{datasets}
