% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_hcl.R
\name{scale_fill_hcl}
\alias{scale_fill_hcl}
\title{Create your own discrete scale}
\usage{
scale_fill_hcl(..., palette = "RdYlBu", reverse = FALSE)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale}}
  \describe{
    \item{\code{scale_name}}{The name of the scale that should be used for error messages
associated with this scale.}
    \item{\code{name}}{The name of the scale. Used as the axis or legend title. If
\code{waiver()}, the default, the name of the scale is taken from the first
mapping used for that aesthetic. If \code{NULL}, the legend title will be
omitted.}
    \item{\code{breaks}}{One of:
\itemize{
\item \code{NULL} for no breaks
\item \code{waiver()} for the default breaks (the scale limits)
\item A character vector of breaks
\item A function that takes the limits as input and returns breaks
as output. Also accepts rlang \link[rlang:as_function]{lambda} function
notation.
}}
    \item{\code{labels}}{One of:
\itemize{
\item \code{NULL} for no labels
\item \code{waiver()} for the default labels computed by the
transformation object
\item A character vector giving labels (must be same length as \code{breaks})
\item A function that takes the breaks as input and returns labels
as output. Also accepts rlang \link[rlang:as_function]{lambda} function
notation.
}}
    \item{\code{limits}}{One of:
\itemize{
\item \code{NULL} to use the default scale values
\item A character vector that defines possible values of the scale and their
order
\item A function that accepts the existing (automatic) values and returns
new ones. Also accepts rlang \link[rlang:as_function]{lambda} function
notation.
}}
    \item{\code{na.translate}}{Unlike continuous scales, discrete scales can easily show
missing values, and do so by default. If you want to remove missing values
from a discrete scale, specify \code{na.translate = FALSE}.}
    \item{\code{na.value}}{If \code{na.translate = TRUE}, what aesthetic value should the
missing values be displayed as? Does not apply to position scales
where \code{NA} is always placed at the far right.}
    \item{\code{drop}}{Should unused factor levels be omitted from the scale?
The default, \code{TRUE}, uses the levels that appear in the data;
\code{FALSE} uses all the levels in the factor.}
    \item{\code{guide}}{A function used to create a guide or its name. See
\code{\link[ggplot2:guides]{guides()}} for more information.}
    \item{\code{super}}{The super class to use for the constructed scale}
  }}

\item{palette}{the name of the palette to generate colors from. A list of all
available palettes can be found by running \code{colorspace::hcl_palettes()}
(default is "RdYlBu").}

\item{reverse}{If \code{TRUE}, the order of the colors is reversed (default is \code{FALSE})'}
}
\value{
Scale layer for the fill aesthetic.
}
\description{
These functions allow you to specify your own set of mappings from levels in the
data to aesthetic values.
}
\details{
The functions `scale_colour_manual()`, `scale_fill_manual()`, `scale_size_manual()`,
etc. work on the aesthetics specified in the scale name: `colour`, `fill`, `size`,
etc. However, the functions `scale_colour_manual()` and `scale_fill_manual()` also
have an optional `aesthetics` argument that can be used to define both `colour` and
`fill` aesthetic mappings via a single function call (see examples). The function
`scale_discrete_manual()` is a generic scale that can work with any aesthetic or set
of aesthetics provided via the `aesthetics` argument.
}
