% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sports_time}
\alias{sports_time}
\title{Peaks times for sports and leisure activities}
\format{
A data frame with 8,092 rows and 3 variables:
\describe{
  \item{activity}{name of the activity}
  \item{time}{time of the day, in hhmm format}
  \item{p}{standardized peak}
}
}
\source{
\url{https://github.com/halhen/viz-pub/blob/master/sports-time-of-day/activity.tsv, https://eagereyes.org/blog/2017/joy-plots}
}
\usage{
sports_time
}
\description{
A dataset containing the peak time for doing 29 sports
and leisure activities.
}
\keyword{datasets}
