% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_horizon.R
\docType{data}
\name{StatHorizon}
\alias{StatHorizon}
\title{Add together two numbers}
\format{
An object of class \code{StatHorizon} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 6.
}
\usage{
StatHorizon
}
\description{
Add together two numbers
}
\keyword{datasets}
