% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posneg.R
\name{posneg}
\alias{posneg}
\title{Internal function}
\usage{
posneg(y, origin, vec)
}
\description{
This function calculates the different positive and
negative boundaries for each data point
based on the origin and the cutpoint vector provided.
}
\keyword{internal}
